requirements.txt
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.1
psycopg2-binary==2.8.6
whitenoise==5.2.0
gunicorn==20.0.4
